tidybot/ai_service/requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1

# AI and ML
torch==2.5.1
torchvision==0.20.1
transformers==4.47.1
Pillow==11.0.0
opencv-python==4.10.0.84
pytesseract==0.3.13
numpy==1.26.4
scikit-learn==1.5.2

# Document processing
nltk==3.9.1
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
chardet==5.2.0
python-magic==0.4.27

# Database and caching
sqlalchemy==2.0.36
aiosqlite==0.20.0
redis==5.2.1
aiocache==0.12.3

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
aiofiles==24.1.0
python-dotenv==1.0.1
pyyaml==6.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1